add block.json and restructure block into separate partials
